+++ crates/common/src/io_ext/mod.rs
@@ -7,8 +7,8 @@ compile_error!("only little-endian architectures are supported");
 
 pub struct CountingReader<R: Read> {
     inner: R,
-    pub offset: u32,
-    pub prev: u32,
+    pub offset: usize,
+    pub prev: usize,
 }
 
 impl<R: Read> CountingReader<R> {
@@ -34,7 +34,7 @@ impl<R: Read> CountingReader<R> {
     pub fn read_exact(&mut self, buf: &mut [u8]) -> Result<()> {
         self.inner.read_exact(buf)?;
         self.prev = self.offset;
-        self.offset += buf.len() as u32;
+        self.offset += buf.len();
         Ok(())
     }
 
@@ -116,7 +116,7 @@ impl<R: Read> CountingReader<R> {
 
 impl<R: Read + Seek> CountingReader<R> {
     #[inline]
-    pub fn seek(&mut self, pos: SeekFrom) -> crate::Result<u32> {
+    pub fn seek(&mut self, pos: SeekFrom) -> crate::Result<usize> {
         let offset = self.inner.seek(pos)?;
         let offset = offset
             .try_into()
