+++ crates/nodes/src/pm/lod/data.rs
@@ -29,7 +29,7 @@ struct LodPmC {
 }
 impl_as_bytes!(LodPmC, 92);
 
-fn assert_lod(lod: &LodPmC, offset: u32) -> Result<(bool, Range)> {
+fn assert_lod(lod: &LodPmC, offset: usize) -> Result<(bool, Range)> {
     trace!("{:#?}", lod);
     let level = assert_that!("level", bool lod.level, offset + 0)?;
