+++ crates/nodes/src/pm/object3d/info.rs
@@ -45,7 +45,7 @@ const VARIABLE_FLAGS: NodeBitFlags = NodeBitFlags::from_bits_truncate(
 #[allow(clippy::collapsible_else_if)]
 pub fn assert_variants(
     node: NodeVariantsPm,
-    offset: u32,
+    offset: usize,
     mesh_index_is_ptr: bool,
 ) -> Result<NodeVariantPm> {
     // cannot assert name
